+++ src/pysight/lst_tools.py
@@ -117,8 +117,14 @@ class Analysis(object):
 
         dict_of_data = {}
         self.data_to_grab = ['abs_time']
-        if self.use_sweeps:
+        if self.use_sweeps: # Sweeps as lines - generate a "fake" line signal
             self.data_to_grab.extend(('edge', 'sweep', 'time_rel_sweep'))
+            sweep_vec = np.arange(df['sweep'].max() + 1, dtype=np.uint64)
+            if len(sweep_vec) < 2:
+                warnings.warn("All data was registered to a single sweep. Line data will be completely simulated.")
+            else:
+                dict_of_data['Lines'] = pd.DataFrame(sweep_vec * self.data_range,
+                                                     columns=['abs_time'], dtype=np.uint64)
 
         for key in self.dict_of_inputs:
             relevant_values = df.loc[df['channel'] == self.dict_of_inputs[key], self.data_to_grab]